Restructure the way tile freeing is handled.  When a tile is removed from the tile cache it's ssg elements are disconnected from the main ssg scene graph, and then the tile is thrown on the end of a delete queue.  The tilemgr->update() routine runs every frame.  It looks at this queue and if it is non-empty, it incrementally frees the compents of the first tile on the queue.  When the tile is completely free it is removed from the queue.
The amount of time to free the memory for even a single tile can be quite
substantial, especially with the increased overhead of dynamic/random
ground objects.  This change allows the system to spread the work of freeing
tile memory out over many frames so you don't get a noticable single frame
hit or stutter.