Removed the FG3DModel class and replaced it with fgLoad3DModel. Animations are now contained within the scene graph itself and are updated whenever the graph is traversed -- that saves time by not updating animations not currently in sight, and it allows animations to be used for static objects and random objects as well.
Added new FGModelLoader and FGTextureLoader classes.  These are intern
tables for models, to guarantee (mostly) that no model is loaded more
than once.  FGTextureLoader is not yet used anywhere, but
FGModelLoader is now in place everywhere that ssgLoad* used to be
used (thus adding the ability to use animations).

In the future, FGModelLoader will add some interesting functionality,
including the ability to reload 3D models on the fly.