Investigating some wierd behavior where the threaded metar fetcher would occasionally cause a large number of valid stations to be flagged as invalid. This *seemed* like a "race condition" type problem because there were some assumptions in the communication between the main process and the threaded loader which if they broke down could lead to this problem.
In the process of removing this ambiguity, I restructured the threaded
(and non-threaded) metar fetching code a bit.  Some of the top level logic
(which Erik politely left untouched) didn't make nearly as much sense in the
context of a threaded metar loader and could have contributed to some of the
wierdness I was seeing.